analyzer: enable hybrid analyzer
 * hybrid analyzer extends dynamic analyzer and composites aiq
   analyzer
 * with analyzer, customer ae/awb 3a results will be applied to
   aiq maunal settings
 * test: put a customized 3a lib under /usr/lib/xcam/, then
   test-device-manager -f YUYV -m dma -d video -p -a hybrid
