+++ xcore/analyzer_loader.cpp
@@ -19,312 +19,15 @@
  */
 
 #include "analyzer_loader.h"
+#include "dynamic_analyzer.h"
 #include "handler_interface.h"
+#include "hybrid_analyzer.h"
+#include "isp_controller.h"
 #include "x3a_result_factory.h"
 #include "x3a_statistics_queue.h"
 #include <dlfcn.h>
 
 namespace XCam {
-
-class DynamicAeHandler;
-class DynamicAwbHandler;
-class DynamicAfHandler;
-class DynamicCommonHandler;
-
-class DynamicAnalyzer
-    : public X3aAnalyzer
-{
-public:
-    DynamicAnalyzer (XCam3ADescription *desc, SmartPtr<AnalyzerLoader> &loader);
-    ~DynamicAnalyzer ();
-
-public:
-    XCamReturn analyze_ae (XCamAeParam &param);
-    XCamReturn analyze_awb (XCamAwbParam &param);
-    XCamReturn analyze_af (XCamAfParam &param);
-
-protected:
-    virtual SmartPtr<AeHandler> create_ae_handler ();
-    virtual SmartPtr<AwbHandler> create_awb_handler ();
-    virtual SmartPtr<AfHandler> create_af_handler ();
-    virtual SmartPtr<CommonHandler> create_common_handler ();
-    virtual XCamReturn internal_init (uint32_t width, uint32_t height, double framerate);
-    virtual XCamReturn internal_deinit ();
-
-    virtual XCamReturn configure_3a ();
-    virtual XCamReturn pre_3a_analyze (SmartPtr<X3aStats> &stats);
-    virtual XCamReturn post_3a_analyze (X3aResultList &results);
-
-    XCamReturn create_context ();
-    void destroy_context ();
-
-private:
-    XCamReturn convert_results (XCam3aResultHead *from[], uint32_t from_count, X3aResultList &to);
-    XCAM_DEAD_COPY (DynamicAnalyzer);
-
-private:
-    XCam3ADescription           *_desc;
-    XCam3AContext               *_context;
-    SmartPtr<X3aStats>           _cur_stats;
-    SmartPtr<DynamicCommonHandler> _common_handler;
-    SmartPtr<AnalyzerLoader>       _loader;
-};
-
-class DynamicAeHandler
-    : public AeHandler
-{
-public:
-    explicit DynamicAeHandler (DynamicAnalyzer *analyzer)
-        : _analyzer (analyzer)
-    {}
-    virtual XCamReturn analyze (X3aResultList &output) {
-        XCAM_UNUSED (output);
-        AnalyzerHandler::HandlerLock lock(this);
-        XCamAeParam param = this->get_params_unlock ();
-        return _analyzer->analyze_ae (param);
-    }
-
-private:
-    DynamicAnalyzer *_analyzer;
-};
-
-class DynamicAwbHandler
-    : public AwbHandler
-{
-public:
-    explicit DynamicAwbHandler (DynamicAnalyzer *analyzer)
-        : _analyzer (analyzer)
-    {}
-    virtual XCamReturn analyze (X3aResultList &output) {
-        XCAM_UNUSED (output);
-        AnalyzerHandler::HandlerLock lock(this);
-        XCamAwbParam param = this->get_params_unlock ();
-        return _analyzer->analyze_awb (param);
-    }
-
-private:
-    DynamicAnalyzer *_analyzer;
-};
-
-class DynamicAfHandler
-    : public AfHandler
-{
-public:
-    explicit DynamicAfHandler (DynamicAnalyzer *analyzer)
-        : _analyzer (analyzer)
-    {}
-    virtual XCamReturn analyze (X3aResultList &output) {
-        XCAM_UNUSED (output);
-        AnalyzerHandler::HandlerLock lock(this);
-        XCamAfParam param = this->get_params_unlock ();
-        return _analyzer->analyze_af (param);
-    }
-
-private:
-    DynamicAnalyzer *_analyzer;
-};
-
-class DynamicCommonHandler
-    : public CommonHandler
-{
-    friend class DynamicAnalyzer;
-public:
-    explicit DynamicCommonHandler (DynamicAnalyzer *analyzer)
-        : _analyzer (analyzer)
-    {}
-    virtual XCamReturn analyze (X3aResultList &output) {
-        XCAM_UNUSED (output);
-        AnalyzerHandler::HandlerLock lock(this);
-        return XCAM_RETURN_NO_ERROR;
-    }
-
-private:
-    DynamicAnalyzer *_analyzer;
-};
-
-DynamicAnalyzer::DynamicAnalyzer (XCam3ADescription *desc, SmartPtr<AnalyzerLoader> &loader)
-    : X3aAnalyzer ("DynamicAnalyzer")
-    , _desc (desc)
-    , _context (NULL)
-    , _loader (loader)
-{
-}
-
-DynamicAnalyzer::~DynamicAnalyzer ()
-{
-    destroy_context ();
-}
-
-XCamReturn
-DynamicAnalyzer::create_context ()
-{
-    XCam3AContext *context = NULL;
-    XCamReturn ret = XCAM_RETURN_NO_ERROR;
-    XCAM_ASSERT (!_context);
-    if ((ret = _desc->create_context (&context)) != XCAM_RETURN_NO_ERROR) {
-        XCAM_LOG_WARNING ("dynamic 3a lib create context failed");
-        return ret;
-    }
-    _context = context;
-    return XCAM_RETURN_NO_ERROR;
-}
-
-void
-DynamicAnalyzer::destroy_context ()
-{
-    if (_context && _desc && _desc->destroy_context) {
-        _desc->destroy_context (_context);
-        _context = NULL;
-    }
-}
-
-XCamReturn
-DynamicAnalyzer::analyze_ae (XCamAeParam &param)
-{
-    XCAM_ASSERT (_context);
-    return _desc->analyze_ae (_context, &param);
-}
-
-XCamReturn
-DynamicAnalyzer::analyze_awb (XCamAwbParam &param)
-{
-    XCAM_ASSERT (_context);
-    return _desc->analyze_awb (_context, &param);
-}
-
-XCamReturn
-DynamicAnalyzer::analyze_af (XCamAfParam &param)
-{
-    XCAM_ASSERT (_context);
-    return _desc->analyze_af (_context, &param);
-}
-
-SmartPtr<AeHandler>
-DynamicAnalyzer::create_ae_handler ()
-{
-    return new DynamicAeHandler (this);
-}
-
-SmartPtr<AwbHandler>
-DynamicAnalyzer::create_awb_handler ()
-{
-    return new DynamicAwbHandler (this);
-}
-
-SmartPtr<AfHandler>
-DynamicAnalyzer::create_af_handler ()
-{
-    return new DynamicAfHandler (this);
-}
-
-SmartPtr<CommonHandler>
-DynamicAnalyzer::create_common_handler ()
-{
-    if (_common_handler.ptr())
-        return _common_handler;
-
-    _common_handler = new DynamicCommonHandler (this);
-    return _common_handler;
-}
-
-XCamReturn
-DynamicAnalyzer::internal_init (uint32_t width, uint32_t height, double framerate)
-{
-    XCAM_UNUSED (width);
-    XCAM_UNUSED (height);
-    XCAM_UNUSED (framerate);
-    return create_context ();
-}
-
-XCamReturn
-DynamicAnalyzer::internal_deinit ()
-{
-    destroy_context ();
-    return XCAM_RETURN_NO_ERROR;
-}
-
-XCamReturn
-DynamicAnalyzer::configure_3a ()
-{
-    uint32_t width = get_width ();
-    uint32_t height = get_height ();
-    double framerate = get_framerate ();
-    XCamReturn ret = XCAM_RETURN_NO_ERROR;
-
-    XCAM_ASSERT (_context);
-
-    ret = _desc->configure_3a (_context, width, height, framerate);
-    XCAM_FAIL_RETURN (WARNING,
-                      ret == XCAM_RETURN_NO_ERROR,
-                      ret,
-                      "dynamic analyzer configure 3a failed");
-    set_manual_brightness(_brightness_level_param);
-
-    return XCAM_RETURN_NO_ERROR;
-}
-XCamReturn
-DynamicAnalyzer::pre_3a_analyze (SmartPtr<X3aStats> &stats)
-{
-    XCamReturn ret = XCAM_RETURN_NO_ERROR;
-    XCamCommonParam common_params = _common_handler->get_params_unlock ();
-
-    XCAM_ASSERT (_context);
-    _cur_stats = stats;
-    ret = _desc->set_3a_stats (_context, stats->get_stats (), stats->get_timestamp ());
-    XCAM_FAIL_RETURN (WARNING,
-                      ret == XCAM_RETURN_NO_ERROR,
-                      ret,
-                      "dynamic analyzer set_3a_stats failed");
-
-    ret = _desc->update_common_params (_context, &common_params);
-    XCAM_FAIL_RETURN (WARNING,
-                      ret == XCAM_RETURN_NO_ERROR,
-                      ret,
-                      "dynamic analyzer update common params failed");
-
-    return XCAM_RETURN_NO_ERROR;
-}
-
-XCamReturn
-DynamicAnalyzer::post_3a_analyze (X3aResultList &results)
-{
-    XCamReturn ret = XCAM_RETURN_NO_ERROR;
-    XCam3aResultHead *res_array[XCAM_3A_LIB_MAX_RESULT_COUNT];
-    uint32_t res_count = 0;
-
-    xcam_mem_clear (res_array);
-    XCAM_ASSERT (_context);
-    ret = _desc->combine_analyze_results (_context, res_array, &res_count);
-    XCAM_FAIL_RETURN (WARNING,
-                      ret == XCAM_RETURN_NO_ERROR,
-                      ret,
-                      "dynamic analyzer combine_analyze_results failed");
-
-    _cur_stats.release ();
-
-    if (res_count) {
-        ret = convert_results (res_array, res_count, results);
-        XCAM_FAIL_RETURN (WARNING,
-                          ret == XCAM_RETURN_NO_ERROR,
-                          ret,
-                          "dynamic analyzer convert_results failed");
-        _desc->free_results (res_array, res_count);
-    }
-
-    return XCAM_RETURN_NO_ERROR;
-}
-
-XCamReturn
-DynamicAnalyzer::convert_results (XCam3aResultHead *from[], uint32_t from_count, X3aResultList &to)
-{
-    for (uint32_t i = 0; i < from_count; ++i) {
-        SmartPtr<X3aResult> standard_res =
-            X3aResultFactory::instance ()->create_3a_result (from[i]);
-        to.push_back (standard_res);
-    }
-
-    return XCAM_RETURN_NO_ERROR;
-}
 
 AnalyzerLoader::AnalyzerLoader (const char *lib_path)
     : _path (NULL)
@@ -341,10 +44,9 @@ AnalyzerLoader::~AnalyzerLoader ()
         xcam_free (_path);
 }
 
-SmartPtr<X3aAnalyzer>
+XCam3ADescription *
 AnalyzerLoader::load_analyzer (SmartPtr<AnalyzerLoader> &self)
 {
-    SmartPtr<X3aAnalyzer> analyzer;
     XCam3ADescription *desc = NULL;
     const char *symbol = XCAM_3A_LIB_DESCRIPTION;
 
@@ -363,10 +65,37 @@ AnalyzerLoader::load_analyzer (SmartPtr<AnalyzerLoader> &self)
     }
 
     XCAM_LOG_DEBUG ("got symbols(%s) from lib(%s)", symbol, XCAM_STR (_path));
+    return desc;
+}
+
+SmartPtr<X3aAnalyzer>
+AnalyzerLoader::load_dynamic_analyzer (SmartPtr<AnalyzerLoader> &self)
+{
+    SmartPtr<X3aAnalyzer> analyzer;
+    XCam3ADescription *desc = load_analyzer (self);
 
     analyzer = new DynamicAnalyzer (desc, self);
     if (!analyzer.ptr ()) {
-        XCAM_LOG_WARNING ("get symbol(%s) from lib:%s failed", symbol, XCAM_STR (_path));
+        XCAM_LOG_WARNING ("create analyzer(%s) from lib:%s failed", XCAM_STR (analyzer->get_name()), XCAM_STR (_path));
+        close_handle ();
+        return NULL;
+    }
+
+    XCAM_LOG_INFO ("analyzer(%s) created from 3a lib(%s)", XCAM_STR (analyzer->get_name()), XCAM_STR (_path));
+    return analyzer;
+}
+
+SmartPtr<X3aAnalyzer>
+AnalyzerLoader::load_hybrid_analyzer (SmartPtr<AnalyzerLoader> &self,
+                                      SmartPtr<IspController> &isp,
+                                      const char *cpf_path)
+{
+    SmartPtr<X3aAnalyzer> analyzer;
+    XCam3ADescription *desc = load_analyzer (self);
+
+    analyzer = new HybridAnalyzer (desc, self, isp, cpf_path);
+    if (!analyzer.ptr ()) {
+        XCAM_LOG_WARNING ("create analyzer(%s) from lib:%s failed", XCAM_STR (analyzer->get_name()), XCAM_STR (_path));
         close_handle ();
         return NULL;
     }
